.requirements.txt
pytk==0.0.2.1
Pillow==11.3.0
pyinstaller==6.14.2